done with listng redux and functionality
