+++ src/Components/Listings/Amenities.js
@@ -1,0 +1,15 @@
+import React from 'react'
+
+function Amenities() {
+    return (
+            <div>
+                <h1>Amenities</h1>
+            </div>
+    )
+}
+
+export default Amenities
+
+
+
+
